+++ src/sys/os/browser/web/event/mod.rs
@@ -9,6 +9,7 @@ mod tests;
 mod key; // WebKeyLocation
 mod kind; // WebEventKind, WebEventMouse, WebEventPointer, WebKeyLocation
 mod pointer; // WebEventMouse, WebEventPointer
+mod wheel; // WebEventWheel
 
 crate::structural_mods! { // _mods
     _mods {
@@ -17,6 +18,7 @@ crate::structural_mods! { // _mods
             key::*,
             kind::*,
             pointer::*,
+            wheel::*,
         };
     }
 }