+++ src/sys/os/browser/web/web_api.js
@@ -66,7 +66,7 @@ export async function initWasm(wasmPath, imports = {}) {
         if (selector === "document") return document;
         return document.querySelector(selector);
     }
-	// Helper function to map event names to `JsEventKind` indices, matching Rust's repr.
+	// Helper function to map event names to `WebEventKind` indices, matching Rust's repr.
 	function get_event_kind(eventName) {
 		const eventMap = {
 			"click": 1,
@@ -78,8 +78,9 @@ export async function initWasm(wasmPath, imports = {}) {
 			"pointerdown": 7,
 			"pointerup": 8,
 			"pointermove": 9,
-			// 10
-			"resize": 11,
+			"wheel": 10,
+			// "": 11,
+			"resize": 12,
 		};
 		return eventMap[eventName] ?? 0; // Default to 255 for unknown events
 	}
@@ -168,7 +169,7 @@ export async function initWasm(wasmPath, imports = {}) {
 			},
 			event_addListenerMouse: (ePtr, eLen, eventPtr, eventLen, callbackPtr) => {
 				const element = get_element(ePtr, eLen)
-				if (!element) return;
+				if (!element) { console.error(`Element not found for mouse listener`); return; }
 				const event = str_decode(eventPtr, eventLen);
 				const callback = (e) => {
 					const button = e.type === "mousemove" ? -1 : e.button; // -1 for no clicks
@@ -183,7 +184,7 @@ export async function initWasm(wasmPath, imports = {}) {
 			},
 			event_addListenerPointer: (ePtr, eLen, eventPtr, eventLen, callbackPtr) => {
 				const element = get_element(ePtr, eLen)
-				if (!element) return;
+				if (!element) { console.error(`Element not found for pointer listener`); return; }
 				const event = str_decode(eventPtr, eventLen);
 				const callback = (e) => {
 					const etype = get_event_kind(e.type);
@@ -193,9 +194,22 @@ export async function initWasm(wasmPath, imports = {}) {
 						e.preventDefault(); // STOP mouse event from firing
 						wasm.exports.wasm_callback_pointer(
 							callbackPtr, e.pointerId, e.clientX, e.clientY, e.pressure,
-							e.tiltX, e.tiltY, e.twist || 0, etype, time_stamp,
-						);
+							e.tiltX, e.tiltY, e.twist || 0, etype, time_stamp);
 					}
+				};
+				api_events._callbacks.set(callbackPtr, callback);
+				element.addEventListener(event, callback);
+			},
+			event_addListenerWheel: (ePtr, eLen, eventPtr, eventLen, callbackPtr) => {
+				const element = get_element(ePtr, eLen);
+				if (!element) { console.error(`Element not found for wheel listener`); return; }
+				const event = str_decode(eventPtr, eventLen);
+				const callback = (e) => {
+						const buttons = e.buttons;
+						const time_stamp = e.timeStamp;
+						wasm.exports.wasm_callback_wheel(
+							callbackPtr, e.clientX, e.clientY, e.deltaX, e.deltaY,
+							buttons, e.deltaMode, time_stamp);
 				};
 				api_events._callbacks.set(callbackPtr, callback);
 				element.addEventListener(event, callback);
@@ -220,7 +234,7 @@ export async function initWasm(wasmPath, imports = {}) {
 			permissions_query: (namePtr, nameLen) => {
 				return navigator.permissions.query({ name: str_decode(namePtr, nameLen) })
 					.then(result => {
-						switch (result.state) { // JsPermissionState::
+						switch (result.state) { // WebPermissionState::
 							case "granted": return 1; // Granted
 							case "prompt": return 0;  // Prompt
 							case "denied": return -1; // Denied